Architecture assist
Added comments and non-logic code to the RecipeBookViewModel and RecipeBookView as a means to get a foot hold on the architecture of the app. @BearXDown will be adding the logic and any additional properties and methods.

RecipeBookModel.java has not been created, but should be. As for now it should have a single save() method and does a System.out.println().