gitignore
Choosing to keep the 'recipez-core/target' directory out of push/pull requests. as it is generated locally by maven during the build process and it changes every time you build the project.

The files matter, but for the sake of us sharing code it does not. We will almost certainly not work directly in the 'recipez-core/target' directory directly. For the sake of ease when searching through the code we will be working in, I chose the keep the directory out for now.