Naming convention change
RecipeDataStoreModel changed to ObserverModel

The ObserverModel is for our Views (New Recipe, Recipe Book, Meal Planner, Grocery List) so when a change is made to the  data store (currently local to the applicaiton in 'resources\data' directory)

The logic in the file is for our observers to perform and update, but not necessarily what the update is. The update logic is local to each View and each View gets to decide what they do for the update.

The update logic should focus on checking the data store for changes.