Remove instruction logic
Added logic for removing instructions.

Logic will need to be updated when editing instructions logic is added. Instructions will likely become a class on its own. Similar to the [Label <--> Input Control] swap in the Ingredients class. For ease of use, User should be able to edit an already existing entry. Not delete the entry entirely to fix it.